+++ docs/source/conf.py
@@ -24,6 +24,7 @@ sys.path.append(__eric_modules_path)
 # -- Project information -----------------------------------------------------
 project = 'eric-sse'
 author = 'Luca Stretti'
+language = 'en'
 
 # The full version, including alpha/beta/rc tags
 # release = '1.0.0'