+++ eric_sse/entities.py
@@ -1,13 +1,15 @@
 import asyncio
 import traceback
 from abc import ABC, abstractmethod
-from threading import Lock
 from typing import AsyncIterable, Any
 
 import eric_sse
 from eric_sse.exception import InvalidListenerException, NoMessagesException, InvalidChannelException
+from eric_sse.listener import MessageQueueListener
 from eric_sse.message import MessageContract, Message
-from eric_sse.queue import Queue, AbstractMessageQueueFactory, InMemoryMessageQueueFactory
+from eric_sse.queues import Queue
+from eric_sse.persistence import ConnectionRepositoryInterface, InMemoryConnectionRepository
+from eric_sse.connection import Connection
 
 logger = eric_sse.get_logger()
 
@@ -15,129 +17,95 @@ MESSAGE_TYPE_CLOSED = '_eric_channel_closed'
 MESSAGE_TYPE_END_OF_STREAM = '_eric_channel_eof'
 MESSAGE_TYPE_INTERNAL_ERROR = '_eric_error'
 
-class MessageQueueListener(ABC):
-    """
-    Base class for listeners.
 
-    Optionally you can override on_message method if you need to inject code at message delivery time.
-    """
-
-    def __init__(self):
-        self.id: str = eric_sse.generate_uuid()
-        self.__is_running: bool = False
-
-    async def start(self) -> None:
-        self.start_sync()
-
-    def start_sync(self) -> None:
-        logger.debug(f"Starting listener {self.id}")
-        self.__is_running = True
-
-    async def is_running(self) -> bool:
-        return self.is_running_sync()
-
-    def is_running_sync(self) -> bool:
-        return self.__is_running
-
-    async def stop(self) -> None:
-        self.stop_sync()
-
-    def stop_sync(self) -> None:
-        logger.debug(f"Stopping listener {self.id}")
-        self.__is_running = False
-
-    def on_message(self, msg: MessageContract) -> None:
-        """Event handler. It executes when a message is delivered to client"""
-        pass
-
-
-class AbstractChannel(ABC):
-    """
-    Base class for channels.
-
-    Provides functionalities for listeners and message delivery management.
-
-    :class:`eric_sse.queue.InMemoryMessageQueueFactory` is the default implementation used for queues_factory
-    see :class:`eric_sse.prefabs.SSEChannel`
-
-    :param int stream_delay_seconds: Wait time in seconds between message delivery.
-
-    :param eric_sse.queue.AbstractMessageQueueFactory queues_factory:
-    """
-    def __init__(
-            self,
-            stream_delay_seconds: int = 0,
-            queues_factory: AbstractMessageQueueFactory | None = None
-    ):
-        self.id: str = eric_sse.generate_uuid()
-        self.stream_delay_seconds = stream_delay_seconds
-
+class _ConnectionManager:
+    """Maintains relationships between listeners and queues"""
+    def __init__(self, channel_id: str, queues_repository: ConnectionRepositoryInterface):
+        self.__channel_id = channel_id
         self.__listeners: dict[str: MessageQueueListener] = {}
         self.__queues: dict[str: Queue] = {}
-        self.__queues_factory = queues_factory if queues_factory else InMemoryMessageQueueFactory()
+        self.__queues_repository = queues_repository
 
-    def _set_queues_factory(self, queues_factory: AbstractMessageQueueFactory):
-        self.__queues_factory = queues_factory
+
+    def load(self):
+        for c in self.__queues_repository.load(self.__channel_id):
+            self.__listeners[c.listener.id] = c.listener
+            self.__queues[c.listener.id] = c.queue
 
     def add_listener(self) -> MessageQueueListener:
-        """Add the default listener"""
         l = MessageQueueListener()
         self.register_listener(l)
         return l
 
-    def register_listener(self, l: MessageQueueListener):
-        """
-        Adds a listener to channel
-        """
-        self.__listeners[l.id] = l
-        self.__queues[l.id] = self.__queues_factory.create()
+    def register_listener(self, listener: MessageQueueListener):
+        self.__listeners[listener.id] = listener
+        self.__queues[listener.id] = self.__queues_repository.create_queue(listener_id=listener.id)
+        self.__queues_repository.persist(
+            channel_id=self.__channel_id,
+            connection=Connection(listener=listener, queue=self.__queues[listener.id])
+        )
 
-    def remove_listener(self, l_id: str):
-        self._get_queue(listener_id=l_id).delete()
-        del self.__queues[l_id]
-        del self.__listeners[l_id]
+    def register_connection(self, listener: MessageQueueListener, queue: Queue):
+        self.__listeners[listener.id] = listener
+        self.__queues[listener.id] = queue
 
-    def deliver_next(self, listener_id: str) -> MessageContract:
-        """
-        Returns next message for given listener id.
 
-        Raises a NoMessagesException if queue is empty
-        """
-        listener = self.get_listener(listener_id)
-        if listener.is_running_sync():
-            msg = self._get_queue(listener_id).pop()
-            listener.on_message(msg)
-            return msg
-        raise NoMessagesException
+    def remove_listener(self, listener_id: str):
+        del self.__queues[listener_id]
+        del self.__listeners[listener_id]
 
-    def _get_queue(self, listener_id: str) -> Queue:
+        self.__queues_repository.delete(channel_id=self.__channel_id, listener_id=listener_id)
+
+    def get_queue(self, listener_id: str) -> Queue:
         try:
             return self.__queues[listener_id]
         except KeyError:
-            raise InvalidListenerException(f"Invalid listener {listener_id}")
-
-    def _set_queue(self, listener_id: str, queue: Queue):
-        self.__queues[listener_id] = queue
-
-    def dispatch(self, listener_id: str, msg: MessageContract):
-        """Adds a message to listener's queue"""
-
-        self.__add_to_queue(listener_id, msg)
-        logger.debug(f"Dispatched {msg} to {listener_id}")
-
-    def __add_to_queue(self, listener_id: str, msg: MessageContract):
-        self._get_queue(listener_id).push(msg)
-
-    def broadcast(self, msg: MessageContract):
-        """Enqueue a message to all listeners"""
-        for listener_id in self.__listeners.keys():
-            self.dispatch(listener_id, msg=msg)
+            raise InvalidListenerException(f"Invalid listener {listener_id}") from None
 
     def get_listener(self, listener_id: str) -> MessageQueueListener:
         try:
             return self.__listeners[listener_id]
         except KeyError:
             raise InvalidListenerException
+
+    def get_listeners(self) -> dict[str, MessageQueueListener]:
+        """Returns a dict mapping listener ids to listeners"""
+        return self.__listeners
+
+class AbstractChannel(ABC):
+    """
+    Base class for channels.
+
+    Provides functionalities for listeners and message delivery management. Channel needs to be started by calling to **open()** method.
+
+    :class:`eric_sse.queue.InMemoryMessageQueueFactory` is the default implementation used for queues_factory
+    see :class:`eric_sse.prefabs.SSEChannel`
+
+    :param int stream_delay_seconds: Wait time in seconds between message delivery.
+
+    :param eric_sse.persistence.ConnectionRepositoryInterface connections_repository:
+    """
+    def __init__(
+            self,
+            channel_id: str | None = None,
+            stream_delay_seconds: int = 0,
+            connections_repository: ConnectionRepositoryInterface | None = None
+    ):
+        self.__id: str = eric_sse.generate_uuid() if channel_id is None else channel_id
+        assert self.__id is not None
+        self.stream_delay_seconds = stream_delay_seconds
+
+        connections_repository = connections_repository if connections_repository else InMemoryConnectionRepository()
+        self.__connection_manager: _ConnectionManager = _ConnectionManager(self.__id, connections_repository)
+
+    @property
+    def id(self) -> str:
+        return self.__id
+
+    def open(self):
+        """Starts service"""
+        self.__connection_manager.load()
+
 
     @abstractmethod
     def adapt(self, msg: MessageContract) -> Any:
@@ -147,12 +115,13 @@ class AbstractChannel(ABC):
         """
         Entry point for message streaming
 
-        A message with type = 'error' is yeld on invalid listener or channel
+        A message with type = 'error' is yield on invalid listener
         """
 
-        def new_messages():
+        async def new_messages():
             try:
-                yield self.deliver_next(listener.id)
+                result = self.deliver_next(listener.id)
+                yield result
             except NoMessagesException:
                 ...
 
@@ -160,12 +129,12 @@ class AbstractChannel(ABC):
 
             while True:
                 # If client closes connection, stop sending events
-                if not await listener.is_running():
+                if not listener.is_running():
                     logger.debug("Listener stopped. Exiting")
                     break
 
                 try:
-                    for message in new_messages():
+                    async for message in new_messages():
                         yield self.adapt(message)
 
                     await asyncio.sleep(self.stream_delay_seconds)
@@ -179,12 +148,61 @@ class AbstractChannel(ABC):
         async for event in event_generator():
             yield event
 
+    def add_listener(self) -> MessageQueueListener:
+        """Add the default listener and creates corresponding queue"""
+        l = MessageQueueListener()
+        self.register_listener(l)
+        return l
+
+    def register_listener(self, listener: MessageQueueListener):
+        """Registers listener and creates corresponding queue"""
+        return self.__connection_manager.register_listener(listener)
+
+    def register_connection(self, listener: MessageQueueListener, queue: Queue):
+        """Registers a Connection with listener and queue"""
+        return self.__connection_manager.register_connection(listener, queue)
+
+    def remove_listener(self, listener_id: str):
+        self.__connection_manager.remove_listener(listener_id)
+
+    def deliver_next(self, listener_id: str) -> MessageContract:
+        """
+        Returns next message for given listener id.
+
+        Raises a NoMessagesException if queue is empty
+        """
+        listener = self.get_listener(listener_id)
+        if listener.is_running():
+            queue = self.__connection_manager.get_queue(listener.id)
+            msg = queue.pop()
+            listener.on_message(msg)
+            return msg
+
+        raise NoMessagesException
+
+    def _get_queue(self, listener_id: str) -> Queue:
+        return self.__connection_manager.get_queue(listener_id)
+
+    def dispatch(self, listener_id: str, msg: MessageContract):
+        """Adds a message to listener's queue"""
+
+        queue = self._get_queue(listener_id)
+        queue.push(msg)
+        logger.debug(f"Dispatched {msg} to {listener_id}")
+
+    def broadcast(self, msg: MessageContract):
+        """Enqueue a message to all listeners"""
+        for listener_id in self.__connection_manager.get_listeners():
+            self.dispatch(listener_id, msg=msg)
+
+    def get_listener(self, listener_id: str) -> MessageQueueListener:
+        return self.__connection_manager.get_listener(listener_id)
+
     async def watch(self) -> AsyncIterable[Any]:
         # TODO tests
         listener = self.add_listener()
-        listener.start_sync()
+        listener.start()
         return self.message_stream(listener)
 
-    def notify_end(self):
-        """Broadcasts a MESSAGE_TYPE_CLOSED Message"""
-        self.broadcast(Message(msg_type=MESSAGE_TYPE_CLOSED))
+    def get_listeners_ids(self) -> list[str]:
+        return [l.id for l in self.__connection_manager.get_listeners().values()]