+++ examples/distributed_app.py
@@ -1,12 +1,11 @@
-import asyncio, logging
+import asyncio
 
 from eric_sse.message import Message, MessageContract
-from eric_sse.prefabs import SimpleDistributedApplicationListener, SSEChannel
+from eric_sse.prefabs import SimpleDistributedApplicationListener, SimpleDistributedApplicationChannel
 
 import  eric_sse
 logger  = eric_sse.get_logger()
 
-ssc = SSEChannel()
 
 def output(m:MessageContract):
     print(m.payload)
@@ -35,29 +34,31 @@ def bye_handler(m: MessageContract) -> list[Message]:
     output(m)
     return close_connection_response()
 
-def create_listener(ch: SSEChannel):
-    l = SimpleDistributedApplicationListener(ch)
+async def create_listener(ch: SimpleDistributedApplicationChannel):
+    l = SimpleDistributedApplicationListener()
     l.set_action('hello', hello_response)
     l.set_action('hello_ack', hello_ack_response)
     l.set_action('bye', bye_handler)
-    l.start_sync()
+    ch.register_listener(l)
+    l.start()
     return l
 
-async def do_stuff(buddy: SimpleDistributedApplicationListener):
+async def do_stuff(buddy: SimpleDistributedApplicationListener, ssc: SimpleDistributedApplicationChannel):
     async for _ in ssc.message_stream(buddy):
         ...
 
 
 async def main():
+    ssc = SimpleDistributedApplicationChannel()
 
-    alice = create_listener(ssc)
-    bob = create_listener(ssc)
+    alice = await create_listener(ssc)
+    bob = await create_listener(ssc)
 
     # Bob says hello to Alice
     bob.dispatch_to(alice, Message(msg_type='hello', msg_payload='hello!'))
 
-    f2 = asyncio.create_task(do_stuff(alice))
-    f1 = asyncio.create_task(do_stuff(bob))
+    f2 = asyncio.create_task(do_stuff(alice, ssc))
+    f1 = asyncio.create_task(do_stuff(bob, ssc))
 
     await f1
     await f2