+++ examples/producer_consumer.py
@@ -4,11 +4,12 @@ from argparse import ArgumentParser
 from random import uniform
 from time import sleep
 from eric_sse import get_logger
-from eric_sse.entities import MessageQueueListener, MESSAGE_TYPE_CLOSED
+from eric_sse.entities import MESSAGE_TYPE_CLOSED
+from eric_sse.listener import MessageQueueListener
 from eric_sse.message import SignedMessage
 from eric_sse.prefabs import DataProcessingChannel
-from eric_redis_queues import RedisQueueFactory
 
+# TODO support to these parameters
 arguments_parser = ArgumentParser()
 arguments_parser.add_argument('-b', choices=['r', 'i'], default='i', help='Backend to use. "r" = redis, "i" = in memory')
 cli_arguments = arguments_parser.parse_args()
@@ -18,7 +19,7 @@ logger = get_logger()
 class Producer:
 
     @staticmethod
-    def produce_num(c: DataProcessingChannel, l: MessageQueueListener, num: int):
+    async def produce_num(c: DataProcessingChannel, l: MessageQueueListener, num: int):
         for i in range(0, num):
             c.dispatch(l.id, SignedMessage(msg_type='counter', msg_payload=i, sender_id='producer'))
         c.dispatch(l.id, SignedMessage(msg_type=MESSAGE_TYPE_CLOSED, sender_id='producer'))
@@ -30,8 +31,7 @@ class Consumer(MessageQueueListener):
         logger.info(f"Received {msg.type}: {msg.payload}")
 
 class MyChannel(DataProcessingChannel):
-    def activate_redis(self):
-        self._set_queues_factory(RedisQueueFactory())
+
     def adapt(self, msg: SignedMessage) -> dict:
         return {'sender_id': msg.sender_id, 'payload': msg.payload}
 
@@ -39,16 +39,13 @@ async def main():
     # Here you can control message deliver frequency and max workers num
     channel = MyChannel(stream_delay_seconds=0, max_workers=6)
 
-    if cli_arguments.b == 'r':
-        logger.info("Activating redis backend")
-        channel.activate_redis()
 
     listener = Consumer()
     channel.register_listener(listener)
 
-    Producer.produce_num(c=channel, l=listener, num=20)
+    await Producer.produce_num(c=channel, l=listener, num=20)
 
-    await listener.start()
+    listener.start()
     async for m in channel.process_queue(listener):
         print(m)