+++ test/mock/listener.py
@@ -1,4 +1,4 @@
-from eric_sse.entities import MessageQueueListener
+from eric_sse.listener import MessageQueueListener
 from eric_sse.message import Message
 
 
@@ -18,4 +18,4 @@ class MessageQueueListenerMock(MessageQueueListener):
             assert msg.payload == self.fixtures[self.num_received].payload
 
         if self.num_received >= self.disconnect_after:
-            self.stop_sync()
+            self.stop()