+++ test/test_sse_channel.py
@@ -2,7 +2,7 @@ import json
 
 from unittest import IsolatedAsyncioTestCase
 from eric_sse.prefabs import SSEChannel
-from eric_sse.entities import MessageQueueListener, Message
+from eric_sse.entities import Message
 from test.mock.listener import MessageQueueListenerMock
 
 
@@ -15,7 +15,7 @@ class SSEStreamTestCase(IsolatedAsyncioTestCase):
         # setup
         channel = self.sut
         listener = channel.add_listener()
-        await listener.start()
+        listener.start()
 
         msg_to_send = Message(msg_type='test', msg_payload={})
         channel.dispatch(listener_id=listener.id, msg=msg_to_send)
@@ -24,14 +24,14 @@ class SSEStreamTestCase(IsolatedAsyncioTestCase):
             self.assertDictEqual(
                 {'data': {}, 'event': 'test', 'retry': channel.retry_timeout_milliseconds}, msg_received
             )
-            await listener.stop()
+            listener.stop()
 
 
     async def test_payload_adapter_json(self):
         self.sut.payload_adapter = json.dumps
         listener = MessageQueueListenerMock(num_messages_before_disconnect=1)
         self.sut.register_listener(listener)
-        listener.start_sync()
+        listener.start()
         self.sut.dispatch(listener.id, Message(msg_type="test", msg_payload={'a': 1}))
 
         async for m in self.sut.message_stream(listener):
@@ -43,15 +43,15 @@ class SSEStreamTestCase(IsolatedAsyncioTestCase):
         self.sut.dispatch(l.id, Message(msg_type='test'))
         self.sut.dispatch(l.id, Message(msg_type='test'))
 
-        l.start_sync()
+        l.start()
         total_messages_received = 0
         async for _ in self.sut.message_stream(listener=l):
             total_messages_received += 1
-            await l.stop()
+            l.stop()
 
-        await l.start()
+        l.start()
         async for _ in self.sut.message_stream(listener=l):
             total_messages_received += 1
-            await l.stop()
+            l.stop()
 
         self.assertEqual(2, total_messages_received)